+++ src/main/java/net/silentchaos512/gear/client/event/TooltipHandler.java
@@ -211,7 +211,12 @@ public final class TooltipHandler {
     private static void getMaterialTraitLines(ItemTooltipEvent event, PartType partType, MaterialInstance material) {
         material.getMaterial().getTraits(partType).forEach(t -> {
             event.getToolTip().add(t.getDisplayName());
-            //event.getToolTip().add(TextUtil.withColor(t.getConditionsText(), TextFormatting.DARK_GRAY));
+
+            // Trait conditions
+            if (event.getFlags().isAdvanced()) {
+                event.getToolTip().add(new StringTextComponent("  ")
+                        .append(TextUtil.withColor(t.getConditionsText(), TextFormatting.DARK_GRAY)));
+            }
         });
     }